feat: WinningInfo 에 LottoNumber 타입의 bonusNumber 필드를 추가
