+++ src/main/kotlin/lotto/LottoGameLauncher.kt
@@ -3,6 +3,7 @@ package lotto
 import lotto.domain.LottoPurchaseInfo
 import lotto.domain.LottoResult
 import lotto.domain.LottoTicket
+import lotto.domain.WinningInfo
 import lotto.service.AutomaticLottoNumberPackagesGenerator
 import lotto.view.input.InputView
 import lotto.view.result.ResultView
@@ -15,7 +16,8 @@ class LottoGameLauncher(private val inputView: InputView, private val resultView
         val ticket = LottoTicket.from(purchaseInfo.purchaseCount, AutomaticLottoNumberPackagesGenerator())
         resultView.showLottoTicketNumber(ticket)
 
-        val winningInfo = inputView.getWinningNumbers()
-        resultView.showResultStatistics(LottoResult.from(ticket, winningInfo, purchaseInfo.purchaseAmount))
+        val winningNumbers = inputView.getWinningNumbers()
+        val bonusNumber = inputView.getBonusNumber(winningNumbers)
+        resultView.showResultStatistics(LottoResult.from(ticket, WinningInfo(winningNumbers, bonusNumber), purchaseInfo.purchaseAmount))
     }
 }
