+++ src/main/kotlin/lotto/view/input/InputView.kt
@@ -3,10 +3,9 @@ package lotto.view.input
 import lotto.domain.LottoNumber
 import lotto.domain.LottoNumberPackage
 import lotto.domain.LottoPurchaseAmount
-import lotto.domain.WinningInfo
 
 interface InputView {
     fun getPurchaseAmount(): LottoPurchaseAmount
-    fun getWinningNumbers(): WinningInfo
+    fun getWinningNumbers(): LottoNumberPackage
     fun getBonusNumber(winningNumbers: LottoNumberPackage): LottoNumber
 }
